+++ components/Dropdown/src/components/dropdown/index.js
@@ -9,9 +9,9 @@ import {
   TouchableWithoutFeedback,
   Dimensions,
   Platform,
-  ViewPropTypes,
   I18nManager,
 } from 'react-native';
+import { TextPropTypes, ViewPropTypes } from 'deprecated-react-native-prop-types';
 import { Ripple } from '../../../../Ripple';
 import { TextField } from '../../../../TextField';
 
@@ -132,7 +132,7 @@ export default class Dropdown extends PureComponent {
     disabledItemColor: PropTypes.string,
     baseColor: PropTypes.string,
 
-    itemTextStyle: Text.propTypes.style,
+    itemTextStyle: TextPropTypes.style,
 
     itemCount: PropTypes.number,
     itemPadding: PropTypes.number,