+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.1.1
+- Fix issues with deprecated text and view proptypes by using `deprecated-react-native-prop-types`
+- Allow 0 values as feedback - fixes an issue where 0 scores in NPS blocks weren't posted as feedback
+- Update API URL for fetching form configuration
+- Remove unused function and testing URL
 # 2.1.0
 - Add Mopinion Metrics
 - Webform autoclose